feat: Separate TransactionRequest and TransactionRequestBuilder
