+++ tests/integration/main.rs
@@ -8,7 +8,8 @@ use miden_client::{
     },
     sync::NoteTagSource,
     transactions::{
-        PaymentTransactionData, TransactionExecutorError, TransactionRequest, TransactionStatus,
+        PaymentTransactionData, TransactionExecutorError, TransactionRequestBuilder,
+        TransactionStatus,
     },
     ClientError,
 };
@@ -36,13 +37,14 @@ async fn test_added_notes() {
     // Mint some asset for an account not tracked by the client. It should not be stored as an
     // input note afterwards since it is not being tracked by the client
     let fungible_asset = FungibleAsset::new(faucet_account_header.id(), MINT_AMOUNT).unwrap();
-    let tx_request = TransactionRequest::mint_fungible_asset(
+    let tx_request = TransactionRequestBuilder::mint_fungible_asset(
         fungible_asset,
         AccountId::try_from(ACCOUNT_ID_REGULAR).unwrap(),
         NoteType::Private,
         client.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
     println!("Running Mint tx...");
     execute_tx_and_sync(&mut client, faucet_account_header.id(), tx_request).await;
 
@@ -71,20 +73,22 @@ async fn test_multiple_tx_on_same_block() {
 
     // Do a transfer from first account to second account
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
-    let tx_request_1 = TransactionRequest::pay_to_id(
+    let tx_request_1 = TransactionRequestBuilder::pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         None,
         NoteType::Private,
         client.rng(),
     )
-    .unwrap();
-    let tx_request_2 = TransactionRequest::pay_to_id(
+    .unwrap()
+    .build();
+    let tx_request_2 = TransactionRequestBuilder::pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         None,
         NoteType::Private,
         client.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
 
     println!("Running P2ID tx...");
 
@@ -161,13 +165,15 @@ async fn test_p2id_transfer() {
     // Do a transfer from first account to second account
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
     println!("Running P2ID tx...");
-    let tx_request = TransactionRequest::pay_to_id(
+    let tx_request = TransactionRequestBuilder::pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         None,
         NoteType::Private,
         client.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
+
     let note = tx_request.expected_output_notes().next().unwrap().clone();
     let transaction_id = execute_tx(&mut client, from_account_id, tx_request).await;
 
@@ -196,7 +202,7 @@ async fn test_p2id_transfer() {
 
     // Consume P2ID note
     println!("Consuming Note...");
-    let tx_request = TransactionRequest::consume_notes(vec![notes[0].id()]);
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![notes[0].id()]).build();
     execute_tx_and_sync(&mut client, to_account_id, tx_request).await;
 
     // Ensure we have nothing else to consume
@@ -250,13 +256,14 @@ async fn test_p2id_transfer_failing_not_enough_balance() {
     // Do a transfer from first account to second account
     let asset = FungibleAsset::new(faucet_account_id, MINT_AMOUNT + 1).unwrap();
     println!("Running P2ID tx...");
-    let tx_request = TransactionRequest::pay_to_id(
+    let tx_request = TransactionRequestBuilder::pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         None,
         NoteType::Private,
         client.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
     execute_failing_tx(
         &mut client,
         from_account_id,
@@ -329,13 +336,14 @@ async fn test_p2idr_transfer_consumed_by_target() {
     let current_block_num = client.get_sync_height().await.unwrap();
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
     println!("Running P2IDR tx...");
-    let tx_request = TransactionRequest::pay_to_id(
+    let tx_request = TransactionRequestBuilder::pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         Some(current_block_num + 50),
         NoteType::Private,
         client.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
     execute_tx_and_sync(&mut client, from_account_id, tx_request.clone()).await;
 
     // Check that note is committed for the second account to consume
@@ -346,7 +354,7 @@ async fn test_p2idr_transfer_consumed_by_target() {
     // Make the `to_account_id` consume P2IDR note
     let note_id = tx_request.expected_output_notes().next().unwrap().id();
     println!("Consuming Note...");
-    let tx_request = TransactionRequest::consume_notes(vec![note_id]);
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![note_id]).build();
     execute_tx_and_sync(&mut client, to_account_id, tx_request).await;
     let (regular_account, seed) = client.get_account(from_account_id).await.unwrap();
     // The seed should not be retrieved due to the account not being new
@@ -404,13 +412,14 @@ async fn test_p2idr_transfer_consumed_by_sender() {
     let current_block_num = client.get_sync_height().await.unwrap();
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
     println!("Running P2IDR tx...");
-    let tx_request = TransactionRequest::pay_to_id(
+    let tx_request = TransactionRequestBuilder::pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         Some(current_block_num + 5),
         NoteType::Private,
         client.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
     execute_tx_and_sync(&mut client, from_account_id, tx_request).await;
 
     // Check that note is committed
@@ -420,7 +429,7 @@ async fn test_p2idr_transfer_consumed_by_sender() {
 
     // Check that it's still too early to consume
     println!("Consuming Note (too early)...");
-    let tx_request = TransactionRequest::consume_notes(vec![notes[0].id()]);
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![notes[0].id()]).build();
     let transaction_execution_result = client.new_transaction(from_account_id, tx_request).await;
     assert!(transaction_execution_result.is_err_and(|err| {
         matches!(
@@ -441,7 +450,7 @@ async fn test_p2idr_transfer_consumed_by_sender() {
 
     // Consume the note with the sender account
     println!("Consuming Note...");
-    let tx_request = TransactionRequest::consume_notes(vec![notes[0].id()]);
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![notes[0].id()]).build();
     execute_tx_and_sync(&mut client, from_account_id, tx_request).await;
 
     let (regular_account, seed) = client.get_account(from_account_id).await.unwrap();
@@ -494,13 +503,14 @@ async fn test_get_consumable_notes() {
     // Do a transfer from first account to second account
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
     println!("Running P2IDR tx...");
-    let tx_request = TransactionRequest::pay_to_id(
+    let tx_request = TransactionRequestBuilder::pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         Some(100),
         NoteType::Private,
         client.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
     execute_tx_and_sync(&mut client, from_account_id, tx_request).await;
 
     // Check that note is consumable by both accounts
@@ -557,7 +567,7 @@ async fn test_get_output_notes() {
     // Do a transfer from first account to second account
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
     println!("Running P2ID tx...");
-    let tx_request = TransactionRequest::pay_to_id(
+    let tx_request = TransactionRequestBuilder::pay_to_id(
         PaymentTransactionData::new(
             vec![Asset::Fungible(asset)],
             from_account_id,
@@ -567,7 +577,8 @@ async fn test_get_output_notes() {
         NoteType::Private,
         client.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
 
     let output_note_id = tx_request.expected_output_notes().next().unwrap().id();
 
@@ -598,13 +609,14 @@ async fn test_import_expected_notes() {
 
     wait_for_node(&mut client_2).await;
 
-    let tx_request = TransactionRequest::mint_fungible_asset(
+    let tx_request = TransactionRequestBuilder::mint_fungible_asset(
         FungibleAsset::new(faucet_account.id(), MINT_AMOUNT).unwrap(),
         client_2_account.id(),
         NoteType::Public,
         client_2.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
     let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
     client_2.sync_state().await.unwrap();
 
@@ -631,13 +643,14 @@ async fn test_import_expected_notes() {
     // If client 2 succesfully consumes the note, we confirm we have MMR and block header data
     consume_notes(&mut client_2, client_2_account.id(), &[input_note.try_into().unwrap()]).await;
 
-    let tx_request = TransactionRequest::mint_fungible_asset(
+    let tx_request = TransactionRequestBuilder::mint_fungible_asset(
         FungibleAsset::new(faucet_account.id(), MINT_AMOUNT).unwrap(),
         first_basic_account.id(),
         NoteType::Private,
         client_2.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
     let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
 
     // Import an uncommited note without verification
@@ -684,13 +697,14 @@ async fn test_import_expected_note_uncommitted() {
 
     wait_for_node(&mut client_2).await;
 
-    let tx_request = TransactionRequest::mint_fungible_asset(
+    let tx_request = TransactionRequestBuilder::mint_fungible_asset(
         FungibleAsset::new(faucet_account.id(), MINT_AMOUNT).unwrap(),
         client_2_account.id(),
         NoteType::Public,
         client_1.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
 
     let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
     client_2.sync_state().await.unwrap();
@@ -727,13 +741,14 @@ async fn test_import_expected_notes_from_the_past_as_committed() {
 
     wait_for_node(&mut client_2).await;
 
-    let tx_request = TransactionRequest::mint_fungible_asset(
+    let tx_request = TransactionRequestBuilder::mint_fungible_asset(
         FungibleAsset::new(faucet_account.id(), MINT_AMOUNT).unwrap(),
         first_basic_account.id(),
         NoteType::Public,
         client_1.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
     let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
 
     let block_height_before = client_1.get_sync_height().await.unwrap();
@@ -837,13 +852,14 @@ async fn test_sync_detail_values() {
 
     // Do a transfer with recall from first account to second account
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
-    let tx_request = TransactionRequest::pay_to_id(
+    let tx_request = TransactionRequestBuilder::pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         Some(new_details.block_num + 5),
         NoteType::Public,
         client1.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
     let note_id = tx_request.expected_output_notes().next().unwrap().id();
     execute_tx_and_sync(&mut client1, from_account_id, tx_request).await;
 
@@ -855,7 +871,7 @@ async fn test_sync_detail_values() {
     assert_eq!(new_details.updated_accounts.len(), 0);
 
     // Consume the note with the second account
-    let tx_request = TransactionRequest::consume_notes(vec![note_id]);
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![note_id]).build();
     execute_tx_and_sync(&mut client2, to_account_id, tx_request).await;
 
     // First client sync should have a new nullifier as the note was consumed
@@ -883,13 +899,14 @@ async fn test_multiple_transactions_can_be_committed_in_different_blocks_without
         let fungible_asset = FungibleAsset::new(faucet_account_id, MINT_AMOUNT).unwrap();
 
         println!("Minting Asset");
-        let tx_request = TransactionRequest::mint_fungible_asset(
+        let tx_request = TransactionRequestBuilder::mint_fungible_asset(
             fungible_asset,
             from_account_id,
             NoteType::Private,
             client.rng(),
         )
-        .unwrap();
+        .unwrap()
+        .build();
 
         println!("Executing transaction...");
         let transaction_execution_result =
@@ -909,13 +926,14 @@ async fn test_multiple_transactions_can_be_committed_in_different_blocks_without
         let fungible_asset = FungibleAsset::new(faucet_account_id, MINT_AMOUNT).unwrap();
 
         println!("Minting Asset");
-        let tx_request = TransactionRequest::mint_fungible_asset(
+        let tx_request = TransactionRequestBuilder::mint_fungible_asset(
             fungible_asset,
             from_account_id,
             NoteType::Private,
             client.rng(),
         )
-        .unwrap();
+        .unwrap()
+        .build();
 
         println!("Executing transaction...");
         let transaction_execution_result =
@@ -945,13 +963,14 @@ async fn test_multiple_transactions_can_be_committed_in_different_blocks_without
         let fungible_asset = FungibleAsset::new(faucet_account_id, MINT_AMOUNT).unwrap();
 
         println!("Minting Asset");
-        let tx_request = TransactionRequest::mint_fungible_asset(
+        let tx_request = TransactionRequestBuilder::mint_fungible_asset(
             fungible_asset,
             from_account_id,
             NoteType::Private,
             client.rng(),
         )
-        .unwrap();
+        .unwrap()
+        .build();
 
         println!("Executing transaction...");
         let transaction_execution_result =
@@ -1059,17 +1078,17 @@ async fn test_consume_multiple_expected_notes() {
         expected_notes.partition(|note| client_notes_ids.contains(&note.id()));
 
     // Create and execute transactions
-    let tx_request_1 = TransactionRequest::consume_notes(
+    let tx_request_1 = TransactionRequestBuilder::consume_notes(
         client_owned_notes.iter().map(|note| note.id()).collect(),
     )
-    .with_authenticated_input_notes(client_owned_notes.iter().map(|note| (note.id(), None)));
+    .with_authenticated_input_notes(client_owned_notes.iter().map(|note| (note.id(), None)))
+    .build();
 
-    let tx_request_2 = TransactionRequest::consume_notes(
+    let tx_request_2 = TransactionRequestBuilder::consume_notes(
         unauth_owned_notes.iter().map(|note| note.id()).collect(),
     )
-    .with_unauthenticated_input_notes(
-        unauth_owned_notes.iter().map(|note| ((*note).clone(), None)),
-    );
+    .with_unauthenticated_input_notes(unauth_owned_notes.iter().map(|note| ((*note).clone(), None)))
+    .build();
 
     let tx_id_1 = execute_tx(&mut client, to_account_ids[0], tx_request_1).await;
     let tx_id_2 = execute_tx(&mut unauth_client, to_account_ids[1], tx_request_2).await;
@@ -1138,13 +1157,14 @@ async fn test_import_consumed_note_with_proof() {
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
 
     println!("Running P2IDR tx...");
-    let tx_request = TransactionRequest::pay_to_id(
+    let tx_request = TransactionRequestBuilder::pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         Some(current_block_num),
         NoteType::Private,
         client_1.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
     execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
     let note = client_1
         .get_input_notes(NoteFilter::Committed)
@@ -1157,7 +1177,7 @@ async fn test_import_consumed_note_with_proof() {
     // Consume the note with the sender account
 
     println!("Consuming Note...");
-    let tx_request = TransactionRequest::consume_notes(vec![note.id()]);
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![note.id()]).build();
     execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
 
     // Import the consumed note
@@ -1196,13 +1216,14 @@ async fn test_import_consumed_note_with_id() {
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
 
     println!("Running P2IDR tx...");
-    let tx_request = TransactionRequest::pay_to_id(
+    let tx_request = TransactionRequestBuilder::pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         Some(current_block_num),
         NoteType::Public,
         client_1.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
     execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
     let note = client_1
         .get_input_notes(NoteFilter::Committed)
@@ -1215,7 +1236,7 @@ async fn test_import_consumed_note_with_id() {
     // Consume the note with the sender account
 
     println!("Consuming Note...");
-    let tx_request = TransactionRequest::consume_notes(vec![note.id()]);
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![note.id()]).build();
     execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
     client_2.sync_state().await.unwrap();
 
@@ -1256,13 +1277,14 @@ async fn test_discarded_transaction() {
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
 
     println!("Running P2IDR tx...");
-    let tx_request = TransactionRequest::pay_to_id(
+    let tx_request = TransactionRequestBuilder::pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         Some(current_block_num),
         NoteType::Public,
         client_1.rng(),
     )
-    .unwrap();
+    .unwrap()
+    .build();
 
     execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
     client_2.sync_state().await.unwrap();
@@ -1275,7 +1297,7 @@ async fn test_discarded_transaction() {
         .clone();
 
     println!("Consuming Note...");
-    let tx_request = TransactionRequest::consume_notes(vec![note.id()]);
+    let tx_request = TransactionRequestBuilder::consume_notes(vec![note.id()]).build();
 
     // Consume the note in client 1 but dont submit it to the node
     let tx_result = client_1.new_transaction(from_account_id, tx_request.clone()).await.unwrap();
